+++ src/test/java/io/swagger/codegen/v3/generators/swift/Swift4ModelTest.java
@@ -35,10 +35,10 @@ public class Swift4ModelTest {
 
         final CodegenProperty property1 = cm.vars.get(0);
         Assert.assertEquals(property1.baseName, "id");
-        Assert.assertEquals(property1.datatype, "Int64");
+        Assert.assertEquals(property1.datatype, "Int");
         Assert.assertEquals(property1.name, "_id");
         Assert.assertNull(property1.defaultValue);
-        Assert.assertEquals(property1.baseType, "Int64");
+        Assert.assertEquals(property1.baseType, "Int");
         Assert.assertTrue(getBooleanValue(property1, CodegenConstants.HAS_MORE_EXT_NAME));
         Assert.assertTrue(property1.required);
         Assert.assertTrue(getBooleanValue(property1, CodegenConstants.IS_PRIMITIVE_TYPE_EXT_NAME));
